Vagrant Local Development
This commit adds playbooks and files necessary to do local development
from within Vagrant.

These playbooks start with a fresh Ubuntu 12.04 machine and:
  - Install Ansible
  - Install the Tower nightly
  - Install all components to exactly mirror production, except Apache
  - Install uwsgi and nginx for local development (since Apache lacks a
working auto-reload)

This isn't entirely perfect -- in particular, developing on task code
is probably challenging until I figure out how to get celery not to
read from the install, but it is a very easy way to get 90% of the way
there very, very quickly.