Project SCM type drop down fix
The SCM type for the projects page was broken when we sorted the drop down options by value. There was no value for 'manual' projects. I had to insert the word 'manual' where it was expecting a blank sting and then remove the word manual before saving the job template to the API