Remove open modals when logging in
If the session times out, and the user attempts to open a modal dialog, then it must be closed due to the fact that bootstrap modals do not work with more than one modal open. The login modal is a bootstrap modal so all other modals need to be closed when it is shown.