demodalizing the license modal
the license modal is now its own standalone page. this commit includes demodalizing the license page, directing the user to this page if no license exists, redirects the user to hte home page when a license is entered, and adds the license type to the license page and the About Tower modal. It also includes a small change to the aw-feature directive to better handle a session that does not have a license yet.