Resolve fact versions before looking up fact views
This adds a step to the fact version lookup that attempts to resolve
fact versions before looking up their views. Resolving a fact version
means finding another version to use when there is a problem with the
one the user selected.

In this case, the resolution process works as follows:

1. Check if either request returned an empty list

If so, request again, changing the dates to look back a year. This
will almost always return something.

2. If we are requesting multiple facts for the same host, and both of
the resolved versions have the same timestamp, they are duplicates and
we cannot compare against them.

Instead we instruct it to grab the selected and the previous versions for
comparison.

For host-to-host view, this also updates the timestamp in the header to
reflect the actual timestamps. It _does not_ update the date picker
dates yet.