Treat about partial as a static asset
Now bundling all partials into `templates.js`. This didn't work so well
for the about partial because the cowsay <pre> tag got all messed up
when it was pre-cached in JavaScript. Moving this to assets to show that
it's not meant to be bundled, but downloaded as an ajax request instead.