Fix empty standard out behavior
Previously we could just check for the absence of the stdout file to
know we didn't yet/wouldn't ever have any stdout content.  Now that we
are creating the stdout file ahead of time and persisting it we need to
also check to see if the file contains any data.   This is also
necessary for some unit tests.