Gather pip requirements at package build time
Previously, requirements were gathered prior to the `sdist` target, and
included in the sdist tarball.  As some requirements are compiled, this caused
problems where the compiled dependencies were linked against the wrong
libraries.  This pull-request addresses that by moving the requirements
gathering into the `dh_prep` and `%build` steps of the packaging process.