Session timeout after idle session
This commit is the UI portion of the session timeout feature for 2.4 (not incorporating the session_limit yet). This includes actively ending a session when the user's session time has expired, as well as giving the user a notification one minute prior that their session is about to end. It also includes removing 'session_timeout' from config.js. This is now retrieved from the response headers of each REST call. Each time the user make a rest call the session is extended another 1800 seconds (or whatever amount the user might configure with AUTH_TOKEN_EXPIRATION in local_settings.py)