Job Details--details panel and events summary
This PR consists of the restyling for the Job Details page, specifically for the details panel and for the events summary panel. It includes, but is not limited to changes to the tables, icons, status representations, expendables panels, positioning on screen, the search bars, the toggles, and the pie chart. I also added a CodeMirror instance for the Results panel as a follow up to my last PR.

Remaining items: Responsiveness is the main one, a lot of JS side styling was done on the job detail page to make it fit on one screen. I'll need to have a deeper dive with UX to figure out what we want to do when this page is displayed on smaller browsers. I also have not gotten to the stdout out for htis page.