Added default values to the required job template survey questions
This is really to fix #1445, the problem being that "" does not pass
float validity, and the show/hide was triggering a check which included
the hidden-but-still-there float fields. So this is an easy work around,
and I think it's better UX as well since these values are required to be
filled in in many cases.

Fix survey maker add button after toggling password visibility