Revert required org on notification templates
This reverts the validation of a required Organization on the
Notification Template.   After discussion with jlaska and jladd, I think
this can behave fine in this situation.   It's more like a Credential
then where without an Organization the NT only becomes available to
Super Users and the creator of the NT.