Allow users to post some ignored read only fields to job templates without elevated permissions
Previously, posting these fields would cause us to check and see if the
user has not only admin access to the job template but use access on
anything used by the job template (because the fields wern't white
listed as being ok to modify without comprehensive use access). It's
convenient for the QA team to be allowed include these fields in PUT's,
and since they're ignored, this patch is fairly benign and shouldn't
alter any behavior.

 #3076