activity stream, display name of resource as title
related to #4185

* When activity stream is gone to from a resource edit view, it should:
  * only show activity for the particular object
  * display the name of the resource as a title

This commit patch half-ass does those two things. So what's missing?
That questions is best answered by looking at how the feature behaves in
tower 3.0.3. You'll notice that you are "locked" on the resource you are
viewing the activity stream for. With this commit:
* You aren't locked
* The removal of the resource-id filter (smart-search) is borked.

So what do we need to do?:
* Lock the view to the resource being viewed

How do we need to do it?:
* router params.activity_search.values is a "static" thing. It is used
to tell smart-search to not generate search tags. This is what we want
to do. However, this is a contructor-only parameter. Further, the value
"value" gets passed around to all sorts of other UI contructors that are
hard to get access to when they are needed to try and dynamically set
smart-search ignore search-tags.