Fix issue preventing inventory groups from loading
A conditional was in place to check if a URL included the base path,
if so, it was supposed to use the URL as-is, otherwise it should get
the URL by referencing a property in an object containing default
URLs.

The conditional checked only for 'api/v1'. Since we're at v2, the
check failed eventually resulting in a `replace` call on an
undefined value.  I replaced the conditional to pattern match
api/v*/ instead.