UI - scroll to top in pagination
Currently in pagination, when we switch from one page to another,
the view is stuck at the bottom and is slightly inconvenient to
scroll all the way to top, in potentially a large list of records.
So in order to prevent that, this commit sets automatic scroll to
top while switching between pages, using jQuery's animate and
scrollTop methods.

Signed-off-by: Vismay Golwala <vgolwala@redhat.com>